+++ frontend/techpick/src/features/nodeManagement/ui/pick.css.ts
